Generate OAR score within each cluster and add them to full objects metadata

Generate OAR score within each cluster and add them to full objects metadata

Usage

oar_by_cluster(
  data,
  count.filter = 1,
  blacklisted.genes = NULL,
  suffix = "",
  tolerance = TRUE,
  cores = 1
)

Arguments

data

a clustered Seurat (v5) object.

count.filter

a numeric value indicating the minimum fraction of cells expressing any given gene that will be included in the analysis, default is 1. Values between 0.5 and 2 are recommended.

blacklisted.genes

a character vector with gene names to be excluded from the analysis. Default is empty.

suffix

a string to append to the output variables. Default is empty

tolerance

a boolean or numeric value controlling the tolerance threshold for pattern matching. If set to TRUE, then tolerance is automatically adjusted. Alternatively, user may supply a numeric value indicating the maximum fraction of mismatch between pairs of genes for pattern grouping. Values between 0.01 and 0.05 are recommended.

cores

a numeric value indicating the number of cores to use un parallel processing. Use parallel::detectCores() or parallelly::availableCores() to identify possibilities.

Value

a Seurat object with OAR stats added into meta data

Examples

if (FALSE) { # \dontrun{
pmbcs <- oar_by_cluster(pmbcs)
} # }